+++ dal/src/main/java/com/tzq/dal/mapper/order/OrderInfoMapper.java
@@ -10,7 +10,7 @@ public interface OrderInfoMapper {
 
     int deleteByExample(OrderInfoExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(OrderInfo record);
 
@@ -18,7 +18,7 @@ public interface OrderInfoMapper {
 
     List<OrderInfo> selectByExample(OrderInfoExample example);
 
-    OrderInfo selectByPrimaryKey(Integer id);
+    OrderInfo selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") OrderInfo record, @Param("example") OrderInfoExample example);
