+++ dal/src/main/java/com/tzq/dal/model/order/PassengerInfo.java
@@ -3,7 +3,7 @@ package com.tzq.dal.model.order;
 import java.util.Date;
 
 public class PassengerInfo {
-    private Integer id;
+    private Long id;
 
     private String orderno;
 
@@ -19,7 +19,7 @@ public class PassengerInfo {
 
     private String cardnum;
 
-    private Byte cardissueplace;
+    private Integer cardissueplace;
 
     private Date cardexpired;
 
@@ -31,11 +31,11 @@ public class PassengerInfo {
 
     private Date modifytime;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -95,11 +95,11 @@ public class PassengerInfo {
         this.cardnum = cardnum == null ? null : cardnum.trim();
     }
 
-    public Byte getCardissueplace() {
+    public Integer getCardissueplace() {
         return cardissueplace;
     }
 
-    public void setCardissueplace(Byte cardissueplace) {
+    public void setCardissueplace(Integer cardissueplace) {
         this.cardissueplace = cardissueplace;
     }
