+++ dal/src/main/java/com/tzq/dal/service/RuleMatchingLogService.java
@@ -11,7 +11,7 @@ public interface RuleMatchingLogService {
 
     int deleteByExample(RuleMatchingLogExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(RuleMatchingLog record);
 
@@ -19,7 +19,7 @@ public interface RuleMatchingLogService {
 
     List<RuleMatchingLog> selectByExample(RuleMatchingLogExample example);
 
-    RuleMatchingLog selectByPrimaryKey(Integer id);
+    RuleMatchingLog selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(RuleMatchingLog record, RuleMatchingLogExample example);
